moved all static functions to the ApiClientUtil class
