+++ src/main/java/com/laserfiche/repository/api/clients/impl/ApiClient.java
@@ -31,58 +31,6 @@ public abstract class ApiClient {
                 .getObjectMapper();
     }
 
-    protected static String mergeMaxSizeIntoPrefer(int maxSize, String prefer) {
-        if (maxSize == 0)
-            return prefer;
-        else
-            return prefer == null ? String.format("maxpagesize=%d", maxSize) : String.format("%s; maxpagesize=%d",
-                    prefer, maxSize);
-    }
-
-    protected static Map<String, Object> getParametersWithNonDefaultValue(String[] parameterTypes,
-            String[] parameterNames,
-            Object[] parameterValues) {
-        if (parameterTypes == null || parameterNames == null || parameterValues == null) {
-            throw new IllegalArgumentException("Input cannot be null.");
-        }
-        if (parameterTypes.length != parameterNames.length || parameterNames.length != parameterValues.length) {
-            throw new IllegalArgumentException(
-                    "The arrays for parameter types/names/values should have the same length.");
-        }
-        Map<String, Object> paramKeyValuePairs = new HashMap<>();
-        for (int i = 0; i < parameterValues.length; i++) {
-            if (parameterValues[i] != null && !hasDefaultValue(parameterTypes[i], parameterValues[i])) {
-                List<Object> values = new ArrayList<>();
-                if (parameterValues[i] instanceof Object[]) {
-                    Object[] objects = (Object[]) parameterValues[i];
-                    Collections.addAll(values, objects);
-                } else {
-                    values.add(parameterValues[i]);
-                }
-                if (values.size() == 1) {
-                    paramKeyValuePairs.put(parameterNames[i], values.get(0));
-                } else {
-                    paramKeyValuePairs.put(parameterNames[i], values);
-                }
-            }
-        }
-        return paramKeyValuePairs;
-    }
-
-    private static boolean hasDefaultValue(String type, Object value) {
-        switch (type) {
-            case "int":
-                return value
-                        .toString()
-                        .equals("0");
-            case "boolean":
-                return value
-                        .toString()
-                        .equals("false");
-        }
-        return false;
-    }
-
     protected String toJson(Object object) {
         String json = null;
         try {
@@ -91,119 +39,6 @@ public abstract class ApiClient {
             System.err.println(e);
         }
         return json;
-    }
-
-    protected static Map<String, String> getHeadersMap(Headers headers) {
-        return headers
-                .all()
-                .stream()
-                .collect(Collectors.toMap(Header::getName, Header::getValue));
-    }
-
-    protected static boolean isRetryableStatusCode(int statusCode, HttpMethod requestMethod) {
-        boolean isIdempotent = !requestMethod
-                .toString()
-                .equals("POST");
-        return (statusCode >= 500 || statusCode == 408) && isIdempotent;
-    }
-
-    protected static String getRepositoryEndpoint(String regionDomain) {
-        if (regionDomain == null)
-            throw new IllegalArgumentException("regionDomain is null.");
-        return "https://api." + regionDomain + "/repository";
-    }
-
-    protected static String combineURLs(String baseURL, String relativeURL) {
-        char end = baseURL.charAt(baseURL.length() - 1);
-        char begin = relativeURL.charAt(0);
-        String url;
-
-        if ((end != '/' && begin == '/') || (end == '/' && begin != '/')) {
-            url = baseURL + relativeURL;
-        } else if (begin == '/') {
-            url = baseURL + relativeURL.substring(1);
-        } else {
-            url = baseURL + '/' + relativeURL;
-        }
-        return url;
-    }
-
-    protected static boolean isJsonResponse(HttpResponse<Object> httpResponse) {
-        List<String> ResponseContentType = httpResponse
-                .getHeaders()
-                .get("Content-Type");
-        for (int i = 0; i < ResponseContentType.size(); i++) {
-            if (ResponseContentType
-                    .get(i)
-                    .contains("json")) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    protected static void createApiException(HttpResponse<Object> httpResponse, ProblemDetails problemDetails) {
-        int statusCode = httpResponse.getStatus();
-        Map<String, String> headersMap = getHeadersMap(httpResponse.getHeaders());
-        throw ApiException.create(statusCode, headersMap, problemDetails, null);
-    }
-
-    protected static String beforeSend(String url, Map<String, String> headerParametersWithStringTypeValue,
-            HttpRequestHandler httpRequestHandler) {
-        String requestUrl;
-        Request customRequest = new RequestImpl();
-        BeforeSendResult beforeSendResult = httpRequestHandler.beforeSend(customRequest);
-        String authorizationValue = customRequest
-                .headers()
-                .get("Authorization");
-        if (authorizationValue != null) {
-            headerParametersWithStringTypeValue.put("Authorization", authorizationValue);
-        }
-        if (url.startsWith("http")) {
-            requestUrl = url;
-        } else {
-            String apiBasedAddress = getRepositoryEndpoint(beforeSendResult.getRegionalDomain());
-            requestUrl = combineURLs(apiBasedAddress, url);
-        }
-        return requestUrl;
-    }
-
-    private static boolean shouldThrowException(String requestMethod, String queryStringFields,
-            Collection<?> queryStringFieldList, String url, ApiException err) {
-        if (requestMethod.equals(
-                "HEAD")) {
-            return true;
-        }
-        if (queryStringFields != null && queryStringFieldList != null) {
-            return true;
-        }
-        if (requestMethod.equals(
-                "PATCH")) {
-            return true;
-        }
-        if (requestMethod.equals("PUT") && url.contains("template")) {
-            return true;
-        }
-        if (url.contains(
-                "Task") && err
-                .getProblemDetails()
-                .getTitle()
-                .contains("Error")) {
-            return true;
-        }
-        if (err
-                .getProblemDetails()
-                .getTitle()
-                .equals("Authentication failed (reason: Invalid service principal key. (6-45)).")) {
-            return true;
-        }
-        if (err
-                .getProblemDetails()
-                .getTitle()
-                .equals("The user account name or password is incorrect. [9010]")) {
-            return true;
-        }
-        return false;
     }
 
     protected static <TResponse> TResponse sendRequestParseResponse(
@@ -224,7 +59,7 @@ public abstract class ApiClient {
         String responseJson = null;
         while (retryCount <= maxRetries && shouldRetry) {
             try {
-                String requestUrl = beforeSend(url, headerParametersWithStringTypeValue, httpRequestHandler);
+                String requestUrl = ApiClientUtils.beforeSend(url, headerParametersWithStringTypeValue, httpRequestHandler);
                 final HttpRequestWithBody httpRequestWithBody = httpClient.request(requestMethod, requestUrl);
                 HttpRequest<?> httpRequest = httpRequestWithBody;
                 if (queryParameters != null) {
@@ -261,7 +96,7 @@ public abstract class ApiClient {
                 }
 
                 HttpMethod httpMethod = httpRequest.getHttpMethod();
-                Map<String, String> headersMap = getHeadersMap(httpResponse.getHeaders());
+                Map<String, String> headersMap = ApiClientUtils.getHeadersMap(httpResponse.getHeaders());
                 if (requestMethod.equals("HEAD")) {
                     if (httpResponse.getStatus() == 200) {
                         return (TResponse) httpResponse
@@ -275,8 +110,8 @@ public abstract class ApiClient {
                 }
                 int statusCode = httpResponse.getStatus();
                 shouldRetry = httpRequestHandler.afterSend(
-                        new ResponseImpl((short) statusCode)) || isRetryableStatusCode(statusCode, httpMethod);
-                boolean isJsonResponse = isJsonResponse(httpResponse);
+                        new ResponseImpl((short) statusCode)) || ApiClientUtils.isRetryableStatusCode(statusCode, httpMethod);
+                boolean isJsonResponse = ApiClientUtils.isJsonResponse(httpResponse);
                 if (isJsonResponse) {
                     Object body = httpResponse.getBody();
                     if (body
@@ -305,7 +140,7 @@ public abstract class ApiClient {
                         } catch (Exception e) {
                             throw ApiException.create(httpResponse.getStatus(), headersMap, null, e);
                         }
-                        createApiException(httpResponse, problemDetails);
+                        ApiClientUtils.createApiException(httpResponse, problemDetails);
                     }
                 }
             } catch (Exception err) {
@@ -313,7 +148,7 @@ public abstract class ApiClient {
                         .getClass()
                         .getName()
                         .equals("com.laserfiche.api.client.model.ApiException")) {
-                    if (shouldThrowException(requestMethod, queryStringFields, queryStringFieldList, url,
+                    if (ApiClientUtils.shouldThrowException(requestMethod, queryStringFields, queryStringFieldList, url,
                             (ApiException) err)) {
                         throw err;
                     } else {
